deployment/requirements.txt
streamlit
pandas
seaborn
plotly
matplotlib
pillow
scikit-learn